apps/api/requirements.txt
# Dependencies for the FastAPI API service.
fastapi>=0.109,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
email-validator>=2.2,<3.0
sqlalchemy>=2.0,<3.0
greenlet>=3.1,<4.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
pgvector>=0.2,<1.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
httpx>=0.27,<1.0
redis>=5.0,<6.0
celery[redis]>=5.3,<6.0
asyncpraw>=7.8,<8.0
tenacity>=8.2,<9.0
structlog>=24.1,<25.0
python-multipart>=0.0.9,<1.0
slowapi>=0.1.9,<1.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<6.0
pytest-httpx>=0.30,<1.0
pytest-postgresql>=6.0,<7.0
psycopg[binary]>=3.2,<4.0
pytest-xdist>=3.6,<4.0
freezegun>=1.5,<2.0
ruff>=0.7,<1.0
